# backend/apiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /collection/media:
        get:
            tags:
                - CollectionService
            description: 列出收藏夹中的媒体
            operationId: CollectionService_ListMedia4Collection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListMedia4CollectionResponse'
        post:
            tags:
                - CollectionService
            description: 将媒体添加到收藏夹
            operationId: CollectionService_AddMedia2Collection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.AddMedia2CollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.AddMedia2CollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 从收藏夹中移除媒体
            operationId: CollectionService_RemoveMediaFromCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.RemoveMediaFromCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.RemoveMediaFromCollectionResponse'
...
    /comment/media:
        get:
            tags:
                - CommentService
            description: 列出媒体的评论
            operationId: CommentService_ListComment4Media
            parameters:
                - name: mediaId
                  in: query
                  description: '@gotags: json:"mediaId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListComment4MediaResponse'
    /cover/upload:
        post:
            tags:
                - MediaService
            description: 预注册上传封面
            operationId: MediaService_PreSign4UploadCover
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.PreSign4UploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.PreSign4UploadResponse'
...
    /file/{fileId}/finish:
        post:
            tags:
                - MediaService
            description: 通用确认上传完成
            operationId: MediaService_ReportFinishUpload
            parameters:
                - name: fileId
                  in: path
                  description: |-
                    文件id
                     @gotags: json:"file_id,omitempty,string"
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.ReportFinishUploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ReportFinishUploadResponse'
...
    /media/feed:
        post:
            tags:
                - MediaService
            description: 刷媒体
            operationId: MediaService_FeedShortMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.FeedShortMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.FeedShortMediaResponse'
    /media/finish:
        post:
            tags:
                - MediaService
            description: 确认媒体上传完成
            operationId: MediaService_ReportMediaFinishUpload
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.ReportMediaFinishUploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ReportMediaFinishUploadResponse'
    /media/list:
        post:
            tags:
                - MediaService
            description: 获取当前用户的发布媒体列表
            operationId: MediaService_ListPublishedMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.ListPublishedMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListPublishedMediaResponse'
    /media/upload:
        post:
            tags:
                - MediaService
            description: 预注册上传媒体
            operationId: MediaService_PreSign4UploadMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.PreSign4UploadMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.PreSign4UploadMediaResponse'
    /media/{mediaId}:
        get:
            tags:
                - MediaService
            description: 获取媒体信息
            operationId: MediaService_GetMediaById
            parameters:
                - name: mediaId
                  in: path
                  description: '@gotags: json:"media_id,omitempty,string"'
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.GetMediaByIdResponse'
...
components:
    schemas:
...
        apiService.svapi.AddMedia2CollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
        apiService.svapi.AddMedia2CollectionResponse:
            type: object
            properties: {}
...
        apiService.svapi.CollectionMedia:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
                coverUrl:
                    type: string
                title:
                    type: string
                description:
                    type: string
                mediaUrl:
                    type: string
        apiService.svapi.Comment:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                user:
                    $ref: '#/components/schemas/apiService.svapi.CommentUser'
                replyUser:
                    $ref: '#/components/schemas/apiService.svapi.CommentUser'
                content:
                    type: string
                date:
                    type: string
                likeCount:
                    type: string
                replyCount:
                    type: string
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.Comment'
        apiService.svapi.CommentUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        apiService.svapi.FeedShortMediaRequest:
            type: object
            properties:
                latestTime:
                    type: string
                    description: '@gotags: json:"latest_time,omitempty,string"'
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                feedNum:
                    type: string
                    description: '@gotags: json:"feed_num,omitempty,string"'
            description: 刷媒体请求消息类型
        apiService.svapi.FeedShortMediaResponse:
            type: object
            properties:
                meta:
                    $ref: '#/components/schemas/apiService.svapi.Metadata'
                medias:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.Media'
                nextTime:
                    type: string
                    description: '@gotags: json:"next_time,omitempty,string"'
            description: 刷媒体响应消息类型
...
        apiService.svapi.GetMediaByIdResponse:
            type: object
            properties:
                media:
                    $ref: '#/components/schemas/apiService.svapi.Media'
            description: 获取媒体信息响应消息类型
...
        apiService.svapi.ListComment4MediaResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.Comment'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
...
        apiService.svapi.ListMedia4CollectionResponse:
            type: object
            properties:
                medias:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.CollectionMedia'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
        apiService.svapi.ListPublishedMediaRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationRequest'
            description: 获取当前用户的发布媒体列表请求消息类型
        apiService.svapi.ListPublishedMediaResponse:
            type: object
            properties:
                mediaList:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.Media'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
            description: 获取当前用户的发布媒体列表响应消息类型
...
        apiService.svapi.Media:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                author:
                    $ref: '#/components/schemas/apiService.svapi.MediaAuthor'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favorite_count,omitempty,string"'
                commentCount:
                    type: string
                    description: '@gotags: json:"comment_count,omitempty,string"'
                isFavorite:
                    type: boolean
                title:
                    type: string
        apiService.svapi.MediaAuthor:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
        apiService.svapi.Metadata:
            type: object
            properties:
                bizCode:
                    type: integer
                    format: int32
                message:
                    type: string
                domain:
                    type: string
                reason:
                    type: array
                    items:
                        type: string
        apiService.svapi.PaginationRequest:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
                sort:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.SortField'
        apiService.svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
        apiService.svapi.PreSign4UploadMediaRequest:
            type: object
            properties:
                hash:
                    type: string
                    description: md5 hash
                fileType:
                    type: string
                    description: 文件类型，默认mp4
                size:
                    type: string
                    description: |-
                        文件大小，单位byte
                         @gotags: json:"size,omitempty,string"
                filename:
                    type: string
                    description: 文件名
            description: 预注册上传媒体请求消息类型
        apiService.svapi.PreSign4UploadMediaResponse:
            type: object
            properties:
                url:
                    type: string
                    description: minio上传地址
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 预注册上传媒体响应消息类型
        apiService.svapi.PreSign4UploadRequest:
            type: object
            properties:
                hash:
                    type: string
                    description: md5 hash
                fileType:
                    type: string
                    description: 文件类型，默认mp4
                size:
                    type: string
                    description: |-
                        文件大小，单位byte
                         @gotags: json:"size,omitempty,string"
                filename:
                    type: string
                    description: 文件名
            description: 通用小文件单次上传请求消息类型
        apiService.svapi.PreSign4UploadResponse:
            type: object
            properties:
                url:
                    type: string
                    description: minio上传地址
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 通用小文件单次上传响应消息类型
...
        apiService.svapi.RemoveMediaFromCollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
        apiService.svapi.RemoveMediaFromCollectionResponse:
            type: object
            properties: {}
        apiService.svapi.ReportFinishUploadRequest:
            type: object
            properties:
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 通用确认上传完成请求消息类型
        apiService.svapi.ReportFinishUploadResponse:
            type: object
            properties:
                url:
                    type: string
                    description: url
            description: 通用确认上传完成响应消息类型
        apiService.svapi.ReportMediaFinishUploadRequest:
            type: object
            properties:
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
                title:
                    type: string
                    description: 媒体标题
                coverUrl:
                    type: string
                    description: 媒体封面地址
                description:
                    type: string
                    description: 媒体描述
                mediaUrl:
                    type: string
                    description: 媒体地址
            description: 确认媒体上传完成请求消息类型
        apiService.svapi.ReportMediaFinishUploadResponse:
            type: object
            properties:
                mediaId:
                    type: string
                    description: |-
                        媒体id
                         @gotags: json:"media_id,omitempty,string"
            description: 确认媒体上传完成响应消息类型
...
        apiService.svapi.SortField:
            type: object
            properties:
                field:
                    type: string
                order:
                    type: integer
                    format: enum
...